abstract into hook
